general.css Video Skin Class Stlye Improvement
Previously, the global video element was styled to make it look better on the stand alone page. However; the class `.show-asset-light` is the video skin class assigned to the stand alone page only.

Based on this, I have optimised the style on this class and removed parts no longer needed from the `video-player` class, thereby not affecting any other pages for areas only required on the stand alone page.